Data-adaptive estimation of interactions, effect modification, and mediation using stochastic shift intervention target parameters. In many mixed exposure settings, interactions in the mixture, effect modifiers in the covariates that modify the impact of an exposure and mediating pathways from exposure to outcome are generally unknown. IsoXshift finds these variable sets on one part of the data and estimates counterfactual outcome changes given shifts to exposure on an estimation part of the data. Using cross-validation and targeted learning, estimators are created that utlize machine learning that are unbiased and have the minimum variance.

The IsoXshift function provides an efficient approach to estimate interactions, effect modification, and mediation using targeted minimum loss estimators for counterfactual mean differences under various target parameters. The procedure employs data-adaptive ensemble b-spline models and stochastic interventions, leveraging the sl3 package for ensemble machine learning. The data is split into V folds, in each fold the training data is used to find variable sets using flexible basis function estimators. Given the different variable sets, stochastic intervention target parameters are applied with cross-validated targeted learning.

IsoXshift(
  w,
  a,
  y,
  target_outcome_lvl,
  epsilon,
  estimator = "tmle",
  mu_learner = NULL,
  g_learner = NULL,
  pi_learner = NULL,
  n_folds = 2,
  outcome_type = "continuous",
  parallel = TRUE,
  parallel_type = "multi_session",
  num_cores = 2,
  seed = seed,
  hn_trunc_thresh = 10,
  top_n = 2
)

Arguments

w

A matrix, data.frame, or similar containing a set of baseline covariates. These variables are measured before exposures.

a

matrix, data.frame, or similar containing individual or multiple exposures.

y

numeric vector of observed outcomes.

estimator

The type of estimator to fit: "tmle" for targeted maximum likelihood estimation, or "onestep" for a one-step estimator.

mu_learner

Learners for fitting Super Learner ensembles to the outcome model via sl3.

g_learner

Learners for fitting Super Learner ensembles to the g-mechanism g(A|W) (a probability estimator, not a density estimator) for mediation via sl3.

pi_learner

Learners for fitting Super Learner ensembles to densities via sl3.

n_folds

Number of folds to use in cross-validation, default is 2.

outcome_type

Data type of the outcome, default is "continuous".

parallel

Whether to parallelize across cores (default: TRUE).

parallel_type

Type of parallelization to use if parallel is TRUE: "multi_session" (default), "multicore", or "sequential".

num_cores

Number of CPU cores to use in parallelization (default: 2).

seed

numeric seed value to be passed to all functions.

hn_trunc_thresh

Truncation level for the clever covariate (default: 10).

z

matrix, data.frame, or similar containing individual or multiple mediators (optional).

deltas

A numeric value indicating the shift in exposures to define the target parameter, with respect to the scale of the exposures (A). If adaptive_delta is true, these values will be reduced.

var_sets

A list specifying variable sets for deterministic IsoXshift usage. Example: var_sets <- c("A_1", "A_1-Z_2") where the analyst provides variable sets for exposures, exposure-mediator, or exposure-covariate relationships.

fluctuation

Method used in the targeting step for TML estimation: "standard" or "weighted". This determines where to place the auxiliary covariate in the logistic tilting regression.

e_learner

Learners for fitting Super Learner ensembles to the e-mechanism g(A|Z,W) (a probability estimator, not a density estimator) for mediation via sl3.

zeta_learner

Learners for fitting Super Learner ensembles to the outcome model via sl3..

mediator_type

Data type of the mediator, default is "continuous".

quantile_thresh

Threshold based on quantiles of the F-statistic, used to identify "important" basis functions in the data-adaptive procedure.

verbose

Whether to run verbosely (default: FALSE).

adaptive_delta

If TRUE, reduces the user-specified delta until the Hn calculated for a shift does not have any observation greater than hn_trunc_thresh (default: FALSE).

n_mc_sample

Number of iterations to be used for the Monte Carlo integration procedure when using continuous exposures (default: 1000).

exposure_quantized

Whether the exposure has been discretized into bins, in which case the integration procedure is skipped and weighted sums are used instead (default: FALSE).

mediator_quantized

If the mediator is discretized, a multinomial ML function is used in this regression to avoid density estimation (default: FALSE).

density_type

Type of density estimation to be used: "sl" for Super Learner (default) or "hal" for highly adaptive lasso.

n_bins

Number of bins for quantizing the exposure if mediation is detected (default: 10).

max_degree

Maximum degree of interactions used in the highly adaptive lasso density estimator if used (default: 1).

integration_method

Type of integration to be used in the continuous exposure case: "MC" for Monte Carlo integration (default) or "AQ" for adaptive quadrature.

use_multinomial

Whether to use multinomial regression for binned exposures (default: FALSE).

discover_only

TRUE/FALSE. If TRUE, only the data-adaptive path discovery is done. No estimates are delivered only exposure mediator sets. If FALSE paths are both discovered and estimated.

Value

An S3 object of class IsoXshift containing the results of the procedure to compute a TML or one-step estimate of the counterfactual mean under a modified treatment policy that shifts a continuous-valued exposure by a scalar amount delta. These exposures are data-adaptively identified using the CV-TMLE procedure.