Targeted Minimum Loss Estimate of Counterfactual Mean under Stochastic Shift Intervention

Targeted Minimum Loss Estimate of Counterfactual Mean under Stochastic Shift Intervention

tmle_exposhift(
  data_internal,
  Qn_scaled,
  Qn_unscaled,
  Hn,
  fluctuation = "standard",
  eif_reg_type = c("hal", "glm"),
  y,
  estimator = "tmle"
)

Arguments

data_internal

A data.table constructed internally by a call to expo_shift. This contains most of the data required for computing the targeted minimum loss (TML) estimator.

Qn_scaled

An object providing the value of the outcome evaluated after imposing min max scaling. This object is passed in after being constructed by a call to the internal function for Q estimation.

Qn_unscaled

An object providing the value of the outcome evaluated at the natural value of the treatment that is not scaled. This object is passed in after being constructed by a call to the internal function for Q estimation.

Hn

An object providing values of the auxiliary ("clever") covariate, constructed from the treatment mechanism and required for targeted minimum loss-based estimation. This object should be passed in after being constructed by a call to the clever covariate construction function.

fluctuation

The method to be used in the submodel fluctuation step (targeting step) to compute the TML estimator. The choices are "standard" and "weighted" for where to place the auxiliary covariate in the logistic tilting regression.

eif_reg_type

Whether a flexible nonparametric function ought to be used in the dimension-reduced nuisance regression of the targeting step for the censored data case. By default, the method used is a nonparametric regression based on the Highly Adaptive Lasso. Set this to "glm" to instead use a simple linear regression model. In this step, the efficient influence function (EIF) is regressed against covariates contributing to the censoring mechanism (i.e., EIF ~ V | C = 1).

y

Outcome variable name as a character string.

estimator

A character string specifying the estimator to be used. Currently, only "tmle" is supported.

delta

A numeric value indicating the shift in the treatment to be used in defining the target parameter. This is defined with respect to the scale of the treatment (A).

Value

An S3 object of class txshift containing the results of the procedure to compute a TML estimate of the treatment shift parameter.

Details

This function implements the procedure to construct a targeted minimum loss estimate (TMLE) of the counterfactual mean under a modified treatment policy that incorporates a stochastic shift intervention.